[PCB] Pi Backplane (#12)
A connector board, designed to link a Pi to the backplane. This will have two purposes:

- Testing individual functional units (particularly the PC and ALU)
- Acting as memory for the whole machine

The core of the design are two banks of shift registers (one output, one input), which are connected to the backplane connector itself. The Pi sends data to the output shift registers, and reads from the input shift registers. The pins in the backplane connector are grouped into buses (A, B etc.), and the corresponding output shift registers set to high impedance for the entire bus. The only pins which are isolatable individually are the clock and reset pins.