[Test] Test for Instruction Register (#26)
Tests of the Instruction Register (which also holds the 'Set' instruction). This refactors some code from other tests into `constants.py` and `utils.py`. These moves have not been fully tested.

Also note that there needs to be a delay after completing a RESET, in order to allow the RC delays in the RwR logic to fully reset.